DBSNEW-13673_OperationHistory: DMF

See merge request dbs/ios!3608

+++ swift-vortex/Sources/Services/GetOperationDetailService/ResponseMapper/ResponseMapper+mapGetUINDataResponse.swift
@@ -32,7 +32,7 @@ private extension ResponseMapper {
         
         let account: String?
         let accountTitle: String?
-        let amount: Int?
+        let amount: Decimal?
         let billDate: String?
         let billNumber: String?
         let cityName: String?
